Algorithm: fix extract/is_ methods on the python side.
